util: move log-level into own header
To reduce include spam when some code wants to use just the log-level.
